[New] Added test cases for record list add record menu
